requirements.txt
# AI・機械学習
google-generativeai>=0.3.0
numpy>=1.24.0
pillow>=10.0.0

# ユーティリティ
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# 動画処理
ffmpeg-python>=0.2.0
faster-whisper>=0.10.0

# 開発ツール
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0

# PyQt6
PyQt6>=6.6.1 